requirements.txt
numpy
matplotlib
graphviz
six
tqdm
scipy
visdom
opencv-python
scikit-image
scikit-learn
scikit-fmm
pandas
h5py
jupyter
flask==1.1.1
gunicorn==20.0.4
hydra-core
